+++ src/vm-repair/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 ++++++
 Fixed and updated several vm-repair tests for better coverage. 
 Removed and updated broken image aliases pointing at images that no longer existed. 
+Add `--encrypt-recovery-key` string parameter to `vm repair create` to use recovery key provided by the user to unlock the disk for a confidential VM. 
 
 1.0.8
 ++++++
@@ -18,7 +19,6 @@ az command adjustment
 1.0.6
 ++++++
 Add CLI update wait for ASG to wait for the operation done as the async 2rd operation will cancel the 1st call.
-
 
 1.0.5
 ++++++